Consolidated mergeDuplicateProperty stages into a single mergeAllDuplicates
