+++ lib/Pipeline.js
@@ -3,6 +3,7 @@ var Cesium = require('cesium');
 var Promise = require('bluebird');
 var path = require('path');
 
+var MergeDuplicateProperties = require('./MergeDuplicateProperties');
 var RemoveUnusedProperties = require('./RemoveUnusedProperties');
 var addDefaults = require('./addDefaults');
 var addPipelineExtras = require('./addPipelineExtras');
@@ -16,11 +17,6 @@ var convertDagToTree = require('./convertDagToTree');
 var encodeImages = require('./encodeImages');
 var generateNormals = require('./generateNormals');
 var loadGltfUris = require('./loadGltfUris');
-var mergeDuplicateAccessors = require('./mergeDuplicateAccessors');
-var mergeDuplicateMaterials = require('./mergeDuplicateMaterials');
-var mergeDuplicatePrograms = require('./mergeDuplicatePrograms');
-var mergeDuplicateShaders = require('./mergeDuplicateShaders');
-var mergeDuplicateTechniques = require('./mergeDuplicateTechniques');
 var mergeDuplicateVertices = require('./mergeDuplicateVertices');
 var octEncodeNormals = require('./octEncodeNormals');
 var optimizeForVertexCache = require('./optimizeForVertexCache');
@@ -34,6 +30,7 @@ var writeSource = require('./writeSource');
 var defaultValue = Cesium.defaultValue;
 var defined = Cesium.defined;
 
+var mergeAllDuplicates = MergeDuplicateProperties.mergeAllDuplicates;
 var removeAllUnused = RemoveUnusedProperties.removeAllUnused;
 
 module.exports = Pipeline;
@@ -82,15 +79,12 @@ Pipeline.processJSONWithExtras  = function(gltfWithExtras, options) {
     options = defaultValue(options, {});
     addDefaults(gltfWithExtras, options);
     removeAllUnused(gltfWithExtras);
-    mergeDuplicateShaders(gltfWithExtras);
-    mergeDuplicatePrograms(gltfWithExtras);
-    mergeDuplicateTechniques(gltfWithExtras);
-    mergeDuplicateMaterials(gltfWithExtras);
     // It is generally better to merge the duplicate vertices before merging accessors.
     // Once accessors merge, there is more likely to be overlap in accessor usage between primitives
     // which limits the effectiveness of merging duplicate vertices.
     mergeDuplicateVertices(gltfWithExtras);
-    mergeDuplicateAccessors(gltfWithExtras);
+    mergeAllDuplicates(gltfWithExtras);
+    removeAllUnused(gltfWithExtras);
     removeDuplicatePrimitives(gltfWithExtras);
     convertDagToTree(gltfWithExtras);
     combineNodes(gltfWithExtras);
